Reorganize code.
