+++ huntflow_api_client/models/consts.py
@@ -23,3 +23,24 @@ class VacancyState(str, Enum):
     VACANCY_REQUEST_ATTACH = "VACANCY_REQUEST_ATTACH"
     RESUME = "RESUME"
     CREATED = "CREATED"
+
+
+class FieldType(str, Enum):
+    string = "string"
+    integer = "integer"
+    text = "text"
+    date = "date"
+    select = "select"
+    complex = "complex"
+    contract = "contract"
+    reason = "reason"
+    stoplist = "stoplist"
+    compensation = "compensation"
+    dictionary = "dictionary"
+    income = "income"
+    position_status = "position_status"
+    division = "division"
+    region = "region"
+    url = "url"
+    hidden = "hidden"
+    html = "html"
